Save rawSpec for modules from git
Since npm-package-arg strips a preceding `git+` from urls, and the stripped
urls will cause trouble when loading the package the next time, we have to
make sure that in this case at least we save the rawSpec not the spec.

I don't know enough about the difference between these two to judge whether
we should ALWAYS be returning rawSpec.  But hopefully someone will review my
change and if it passes into master in this form, then it's presumably
correct.