add-remote-git: Add support for git submodules in git remotes
This is a fairly simple approach, which does not leverage the git caching
mechansim to cache submodules.  It also doesn't provide a means to disable
automatic initialization, e.g. via a setting in the .gitmodules file.

This also adds documentation for the git+file protocol, since we have
a test case ensuring that this works.  We have no test case for the
git+rsync protocol, even though npm-package-arg seems to have support
for that as well, so that wasn't added to the documentation.

PR-URL: https://github.com/npm/npm/pull/11094
Credit: @gagern
Reviewed-By: @othiym23
Fixes: #1876