gentlyrm: Don't abort early if npm controlled path is missing
Previously if any of the npm controlled paths had an ENOENT then we would
return fales from `isEverInside`– not an error condition, but a failed match.
What we actually want to do is ONLY do that if ALL of the paths are ENOENT. If
any aren't, then we want to test them normally.

PR-URL: https://github.com/npm/npm/pull/11212
Credit: @iarna
Reviewed-By: @othiym23