Ignore dependency modules for cleaner git status
This changes the .gitignore file to exclude our dependency modules by
default.  Those modules which are already committed to the repository will
remain there, and will not be ignored.

We *could* have omitted the exclusion rules, i.e. ignore all files in
node_modules except those which were explicitely added.  That way we may
however get into trouble if one of these submodules gets updated, either
because only files get added but no existing files get modified, or because
the committer of the update adds the modified files but forgets to add any
new but ignored files.  That's the reason to not ignore all the existing
module directories, so that *any* modification in one of those shows up.