test: fix Windows-specific test issues
This commit is a squash of many other commits. They all act on
individual tests. Some of them have notable messages, which are
preserved below:

test: rewrite verify-no-lifecycle-on-repo to work on windows

On windows, the previous version of this would open a new cmd window, run
echo in it, and then leave it there, interactive, and wait for the window to
be closed.

test: Fix gently-rm-linked-module on windows

We just need to make our link explicitly be a junction.  This will work as
long as you run tests on NTFS.  It does NOT work on, for instance, vbox
shared folders.

test: Fix gently-rm-symlinked-global-dir to work on windows

To do this we needed to do three things:
1. Be aware that npm doesn't use the same global path structure on
windows. (No lib, bin, etc, all flat.)
2. Not tie the tests to platform-specific unicode support
3. Create our links as junctions, so ordinary users can make them

test: Rewrite lifecycle-path to use `npm run env`

This test predates the existence of `npm run env`-- by changing to use
it we both get to get rid of an unneeded fixture and fix things to
work on windows.

test: fix config-edit on windows

`child_process.spawn`, which is what the `editor` package uses
does not support running straight-up scripts on Windows. Only `.exe`s

For this reason, it's better to just explicitly use `node` here.

test: Fix publish-access-unscoped-restricted-fails on windows

This involved switching from a direct integration test to using common.npm,
as the former method wasn't resulting in all files being closed which meant
that `rimraf` of the test dir couldn't finish on Windows.

test: Don't Object.keys on non typeof == object

`Object.keys` is valid on any non-null, not undefined value in current
versions of Node.js, but older versions would throw an error for any
non typeof == object.

test: fix gently-rm-linked-module when run from `npm run tap`

`npm run tap` sets the npm_config_prefix env var, which was
overriding the test's own attempt at setting the prefix (via the
PREFIX env var).

test: fix race in cache-shasum by using async rimraf

Credit: @zkat
Credit: @iarna
PR-URL: https://github.com/npm/npm/pull/11444
Reviewed-By: @zkat
Reviewed-By: @iarna
Reviewed-By: @othiym23