doc: Document the --registry flag for `npm search`
We use an internal private module server and I was trying to search the
public registry. Tried this and it worked. I looked for a default CLI
page that might be the better place to indicate which flags (if there
are others?) work with all CLI commands, but the docs aren't laid out
that way.

PR-URL: https://github.com/npm/npm/pull/11490
Credit: @plumlee
Reviewed-By: @iarna