deps: dedupe deeply nested bundled modules
The current tree includes a bunch of nested modules which are causing
the CI build for Node.js to not work on windows.

A bit of forensics found that the tree in
`npm/node_modules/npm-install-check/node_modules/npmlog/node_modules/gauge`
is much larger in the latest release due to duplicated modules.

This commit used npm@3 to dedupe node_modules/npm-install-check
resulting in a much shallower tree.

`npm ls` still runs cleanly