test: Fix npm-self-install test to install to /tmp
Previously this was installing into a temp subdir in test/tap, which
wouldn't catch the case where a module was installed in the local
node_modules folder but not in dependencies, as node would look up
the tree and use the copy from the version of npm being tested.

This solves that by moving the install folder to /tmp (which is where
it was in the original legacy test version).