generate a consistent shasum for git dependencies
Formerly, the shasum was generated from a tar.gz file which was
generated from a cloned git tree. Because git doesn't keep track of file
metadata such as timestamps, the tar.gz file generated from the cloned
git tree has a different shasum during every npm install.

So instead of generating a shasum in this way, we use the sha1 from the
git tree where the source was obtained. The result is a shasum that is
always the same for a given git tree. This allows systems like docker
build to make use of the build cache for build steps that involve
running `npm install` with git-based dependencies.