deprecate: default to `*` instead of `latest`
When a semver spec is not supplied, npa supplies a default spec of
'latest'. This is usually correct, but for deprecation the default has
been to apply to all versions of a package.

This patch just checks the raw spec and if it's empty it overrides with
with the '*' version range, which is what people in the issue discussion
were doing anyway.

Fixes: https://github.com/npm/npm/issues/10974
PR-URL: https://github.com/npm/npm/pull/12079
Credit: @polm
Reviewed-By: @zkat