Improve handling of special cases for files included in packages
Most notably, the package 'history' will not longer be included in the
packages of anything which depends on it. Likewise with packages named
changelog or license.

This was fixed in fstream-npm@1.0.6 from via d82ff81403, but for some
reason the internal tar code duplicated this logic. Removing the
duplication passes all the tests, so seems prudent.

While fixing this bug, it was noticed that the `files-and-ignores` tests
were giving false positives because they use `npm install` instead of
only `npm pack`, the tests have been modified to avoid this.

See also #9642, #10445, #11995