deps: Eliminate use of npa for dependency resolution
Previously we would match against realize-package-specifier sometimes, npa
other times and in the case of file deps this would (of course) result in
inconsistent results.

This patch does two things, first it replaces all of the npa calls with
realize-package-specifier and makes those functions callback based. Second,
it caches the results from realize-package-specifier so that we don't have
to keep hitting the disk for the same resolution.

Credit: @iarna
PR-URL: https://github.com/npm/npm/pull/11444
Reviewed-By: @zkat
Reviewed-By: @iarna
Reviewed-By: @othiym23