Further fix to #12347, more reliable _from checks
60da61862885fa904afba7d121860b4282a5b0df provided an initial fix to
shrinkwrap some bugs, but when validating shrink-wrap data, _from
should be compared to rawSpec.

From @iarna: rawSpec is the version part of package@version before it
undergoes normalization. By contrast, raw is the entire specifier. The
existing check worked when the entire specifier was just a URL to a
tarball, but it didn't work for tags like latest.