deps: Rewrite how recalculateMetadata follows dependencies
We keep track of missing dependencies and missing dev dependencies
separately. This allows things like `npm ls` to make decisions about what
it wants to warn about.

Despite having concrete lists of dependencies and dev dependencies we were
guessing as to which was which. This _usually_ works fine, but in cases where
dependencies and dev dependencies share a requirement but not precisely the
same version, it could lead to misleading error messages.

Further, things in the shrinkwrap that are not in the package.json were not being
reported at all.

This change does two things:

1. It rewrites `markDeps` to take an object with name, spec & kind (`prod`
   or `dev`). Previously it took a string of `name@spec` and extracted name
   & spec from that.
2. When a dependency is both a prod and a dev dep, it will be correctly
   categorized. Previously if the dev spec didn't match it would still be
   flagged as a prod mismatch.