deps: Rewrite replaceModule to match by name _or_ path
Previously replaceModule only matched via name, which was appropriate for
updating `requires` and children but was not appropriate for `requiredBy`.
Because `requiredBy` is a list of modules that require this module it can and
will include modules that exist more than once in the physical tree but the
current node as a dependency.  As such, those modules need to be uniqued on
path instead.  We had `userRequired` & `existing` properties on nodes,
indicating that they were explicitly requested for this install or were on
already disk when started, respectively.  But we didn't always use them as
they were added later, sometimes we used huristics based on `_requiredBy`
which was super ugly.

Credit: @iarna
PR-URL: https://github.com/npm/npm/pull/12775