install: loadDeps should skip nodes marked as removed
You might wonder how it's possible for "removed" nodes to get passed to
loadDeps in the first place:

Siblings of a failed dep are unbound from the tree but have a half life, in
the form of the list of nodes that loadDeps is iterating over.  Modules in
requiredBy need to be uniqued on path, as it's totally valid to have more
than one module with the same name requiring the same module.

Credit: @iarna
PR-URL: https://github.com/npm/npm/pull/12775