addRemoteTarball: reuse cached packages
When adding a package via a URL pointing to a tarball, we can try to use a
cached version by attempting to extract the version from the URL and performing
a cache lookup using the name and version. If the cached package was resolved
using the exact same URL then we can reuse it. This makes it safe for packages
added from other sources, or with unusual URL formats from which we cannot
extract the version.