npm: Only enable progressbar if we have a tty
With the older version of gauge (and thus the older version of npmlog), if
you didn't have a tty then `gauge.enable` was ignored.  This is no longer
the case, so now we need to do tty detection before enabling it ourselves.

As a bonus, this now means tty detection is done exactly the same way for
the progress bar as it is for color.

PR-URL: https://github.com/npm/npm/pull/13075
Credit: @iarna