install: Load shrinkwrap before arg metadata
If you run `npm install modulename` it should, if a `npm-shrinkwrap.json` is
present, use the verison found there.  If not, use the version found in your
`package.json` and failing that use `latest`.

Without this patch the first check was being bypassed because version resolution
was being done prior to loading the shrinkwrap, and so checks to match the shrinkwrap
version couldn't succeed.