run-script: Add option to ignore the current node executable for PATH
Add a `--scripts-prepend-node-path` option that controls `npm`’s
behaviour with respect to adding the directory of
the current `node` executable to the `PATH`, even if that means
that `npm` will invoke a different `node` executable than the
one which it is running.

This is a non-invasive solution to some of the problems described in
https://github.com/npm/npm/issues/12318. It should not come with
any drawbacks for npm’s side, as it shifts the responsibility of
making sure that the `PATH` variable contains everything it should
to whoever invoked npm with `--scripts-prepend-node-path=false`.

While https://github.com/npm/npm/pull/12968 may have addressed some
of the problems described there, it does not help in the case that
the `PATH` has explicitly been overridden with the intention of
having a different `node` executed.