run-script: Add option to ignore the current node executable for PATH
Add a `--scripts-prepend-node-path` option that controls `npm`’s
behaviour with respect to adding the directory of
the current `node` executable to the `PATH`, even if that means
that `npm` will invoke a different `node` executable than the
one which it is running.

This is a non-invasive solution to some of the problems described in
https://github.com/npm/npm/issues/12318. It should not come with
any drawbacks for npm’s side, as it shifts the responsibility of
making sure that the `PATH` variable contains everything it should
to whoever invoked npm with `--scripts-prepend-node-path=false`.

While https://github.com/npm/npm/pull/12968 may have addressed some
of the problems described there, it does not help in the case that
the `PATH` has explicitly been overridden with the intention of
having a different `node` executed.

[squash] test: anticipate coverage tooling possibly setting --scripts-prepend-node-path

As one of the motivations for introducing `--scripts-prepend-node-path`
is that `spawn-wrap` (which npm itself uses for coverage) can make use
of that option, anticipate that change in order to not break the
test suite when updating that dependency at a later point.

[squash] tests for scripts-prepend-node-path=auto

[squash] implement `warn-only` setting for `--scripts-prepend-node-path`

Adds a setting that is basically equivalent to `--scripts-prepend-node-path=false`
but emits a warning informing the user about the existence of the
option in the cases in which `npm` currently modifies `PATH`,
as sometimes that’s in the user’s interest.

[squash] fixup tests for windows

PR-URL: https://github.com/npm/npm/pull/13409
Credit: @addaleax
Reviewed-By: @zkat
Reviewed-By: @othiym23