lifecycle: split node_module path parts properly
In lifecycle scripts, any `node_modules/.bin` existing in the heirarchy
should be turned into an entry in the PATH environment variable.
However, prior to this commit, it was splitting based on the string
`node_modules`, rather than restricting it to only path portions like
`/node_modules/` or `\node_modules\`.  So, a path containing an entry
like `my_node_modules` would be improperly split.

Fixes: #13456
PR-URL: https://github.com/npm/npm/pull/13518
Credit: @isaacs
Reviewed-By: @zkat