config: pretend login is a toplevel command
The `login` command has long been an alias for `adduser`.

At the same time, there is an expectation not just of that
particular word being something to look for, but of there being
clear symmetry with `logout`.

So it was a bit confusing when `login` didn't show up in
`npm help` on a technicality. This seems like an acceptable
exception to the rule that says "no aliases in `npm help`".

Fixes: https://github.com/npm/npm/issues/13581
PR-URL: https://github.com/npm/npm/pull/13580
Credit: @zkat