Added sanitation of the maintainers field.
Some (at least Artifactory) doesn't sanitize non-conforming package.json content, and serves it back verbatim. There exists many? packages where the maintainers property is a string instead of an array. This pollutes the NPM cache (when using Artifactory) and breaks the search.js/stripData() function where it tries to do a map over the maintainers property (or an empty array).