git: Allow shallow clones making some installs faster
Adds new "git-shallow-clone" configuration, that is off by default since
it requires Git 1.9 (released February 2014).

Some git repositories are huge, npm only needs the head of each ref it
wants to install.  To make sure it can still potentially reuse the
repository for different branches/tags and also update itself, it will
always 'git fetch origin <the_ref>' which makes sure that single ref is
available and the newest version.

If the ref is a SHA, however, npm will fall back to fetching the entire
repository again, because we can't know where said SHA is.  Some
upstream servers allow you to fetch a SHA, but we can't rely on it.