view: npm show --json only returns one item even when mult match
In npm@2 multiple results were printed with filenames inbetween, resulting
in a stream of plaintext mixed with JSON. This was pretty undesirable.

Previously in, npm@3 the "best" version from the list was picked and shown,
with no facility for getting multiple results.

This changes npm@3 to return an array of matching results.

Fixes: #14030
Credit: @roblg
Reviewed-By: @iarna
PR-URL: https://github.com/npm/npm/pull/14032