npm: remove some ancient aliases
This was introduced in 2011 along with a variety of easter eggs
that made it into the CLI in its early days. Over the years, it was
simply forgotten. This is not the sort of thing the CLI team thinks
is acceptable and does not reflect our values or culture.

Credit: @zkat
PR-URL: https://github.com/npm/npm/pull/14244
Reviewed-By: @othiym23
Reviewed-By: @isaacs