Sync up the dispute policy with the one on npmjs.com
The dispute policy at https://www.npmjs.com/policies/disputes is the
current version of the dispute policy. This syncs up the dispute policy
document here to match what's on the site.

Fixes #13302.

Not sure if the license should be included here or not. Very willing to
update/remove that here.