ci: Node 6 is LTS, 0.10 & 0.12 are unsupported
We still love u, 0.10, but we gotta keep up with the times. This is
_not_ a license to go wild with ES6isms. npm still _works_ (most of the
time) back to 0.8, and that's in large part because of the project's
conservatism about sticking to plain ES5 in both its own code base and
its npm-controlled dependencies. Dependencies not under the team's
control are already hard-deprecating 0.10, though, and the build matrix
already leads to long Travis times.

Credit: @othiym23
Blame: @othiym23
PR-URL: https://github.com/npm/npm/pull/14503
Reviewed-By: @zkat
Reviewed-By: @iarna