wip: Fix preinstall in a systemic way:
The problem it was having was that it was running `preinstall` with a
computing staging folder, but for the top level module and for bundled
dependencies, this wasn't the correct path. This corrects that
by keeping track of the actual current location of the module on disk
in `extract` and `finalize`.