write-file-atomic@1.3.1
When the target is a symlink, write-file-atomic now overwrites the
_destination_ of the symlink, instead of replacing the symlink itself.  This
makes it's behavior match `fs.writeFile`.

Fixed a bug where it would ALWAYS fs.stat to look up default mode and chown
values even if you'd passed them in.  (It still used the values you passed
in, but did a needless stat.)

Fixes: #10223
Fixes: https://github.com/npm/write-file-atomic/issues/5
Credit: @iarna