extract: Rewrite for greater clarity
Specifically, this involves moving things around structurally to do the same
thing d0c6d194 did, but in a more comprehensive manner.

This also fixes a long standing bug where errors from the move step would be
eaten during this phase and as a result we would get mysterious crashes in
the finalize phase when finalize tried to act on them.

Credit: @iarna