Fix wrong symlink direction in link documentation
In npm-link documentation it is written that after running
```npm-link package-name``` symlink is created from global
folder to node_modules of the current project.

However the symlink is created from node_modules of the current
folder to the global folder, otherwise there would be no
way to achieve linked module. So I swapped the direction
in the documentation.