Revert "install/deps: Stop flattening modules that require peer deps"
This reverts commit 06c154fd653d18725d2e760ba825d43cdd807420.

This caused problems because folks were using peer deps to indicate that the
target of the peer dep needed to be able to require the dependency and had
been relying on the fact that peer deps didn't change the shape of the tree
(as of npm@3).

The fix that will actually work for people is for a peer dep to insist on
never being installed deeper than the the thing it relies on.  At the moment
this is tricky because the thing the peer dep relies on may not yet have
been added to the tree, so we don't know where it is.