Fix bug in chunk boundary handling
Somehow when I authored this I went out of my way to handle `\r` on the end of one chunk followed by `\n` on the beginning of the next, yet forgot to take `\r` out of the chunk that it ends (even though it was being added to the beginning, which would double `\r`s on chunk boundaries that aren't followed by `\n`). This change should be fully correct instead -- remove `\r` from the end, add it to the beginning of the next chunk when it was removed from the previous, and after all chunks are done spit out a final `\r` if the last chunk ended in `\r`.