doc: add "notice" loglevel and correct the default
The "notice" loglevel was added and made the default in `npm@5`, but the
docs were not updated to this change. This patch updates the loglevel
docs to the new default and adds "notice" to the list of possible
values.

I don't know how to describe "notice" so I just left it at that.

PR-URL: https://github.com/npm/npm/pull/18397
Credit: @KenanY
Reviewed-By: @iarna