install: fix bug with rollback of failed optional packages on Windows
This change ensures an fsync after updating `package.json` on Windows,
which prevents an issue that intermittently occurs when rolling back a
package with many dependencies such as fsevents.

In testing it does not appear that fsync has as much of a performance
impact on Windows as it does on other platforms, so this should not affect
install times on Windows materially.

Fixes: #17671