install: support legacy npa output for type checking (#18881)
Current npa will pass `registry: true` for specs that will ultimately be
fetched from a registry. Older npa didn't do this.

When dealing with a `node_modules` that was created with older versions of
npm (and thus older versions of npa) we need to gracefully handle these
older spec entries. Failing to do so results in us treating those packages
as if they were http remote deps, which results in invalid lock files.

PR-URL: https://github.com/npm/npm/pull/18881
Credit: @iarna
Reviewed-By: @zkat