install: Handle packages without install metadata gracefully
Currently the only way to get these is to use yarn.

What this means is that we:

Leave them untouched if you didn't as us to do a full install.

Upgrade their metadata if you DID ask us to do a full install.

While: Not messing with a lock file that only has version fields except for
modules we touched.