save/shrinkwrap/version: consolidate code for stringifying package data and add tests (#19907)
This reduces code duplication for detecting indentation and new lines in
`package[-lock].json` and stringifying the data. This also improves
performance on Unix systems since it avoids replacing newlines when
it's not necessary.

Tests have been added to test that `package[-lock].json` files are
written to disk with their original line endings.

PR-URL: https://github.com/npm/npm/pull/19907
Credit: @nwoltman
Reviewed-By: @zkat