install: Let options override global config, always
This introduces the `configGet` function which will prefer results from an
option object over the global config object.  It then introduces is to the
installer and related commands.  Not every use case is covered as yet, but
many are.  This also resulted in a change in the interface to the installer
class, such that it's instantiated with an options object instead of taking
things like `where` as positional arguments.