Fix root user node-gyp permission issues
Per issue #17346, when running npm in docker builds some packages are failing to run installation scripts when installing globally due to the global node_modules directory being owned by root but npm is running the scripts as nobody. This commit attempts to massage the default user configuration to allow executing as root.