audit: ignore devDependencies in production
Having checked how it's done elsewhere in the codebase, this should ignore devDependencies in production.

Aside: This code block is duplicated in several places so I feel like I should be refactoring it into a separate function, but I'm not close enough to the code to know what side-effects that might have.

Fixes #20564